feat(parity): Major parity improvements across Player, Team, INI, W3D systems
Player class (player.rs):
- Added ~45 new fields matching C++ Player class
- Implemented init(), update(), xfer(), crc() methods
- Added PlayerRelationMap integration
- Implemented radar, battle plans, sciences, rank/skill point systems

Team class (team.rs):
- Added TeamTemplateInfo struct with 34 fields
- Implemented TCreateUnitsInfo struct
- Added AttitudeType enum
- Implemented update_state(), kill_team(), evacuate_team()
- Added trigger area detection methods
- Implemented complete xfer() for save/load
- Added TeamPrototype methods

INI Parsers:
- Added 5 new parsers: InGameUI, CommandMap, HeaderTemplate, ScriptAction, ScriptCondition
- Fixed GameLOD wrapper functions for type conversion
- Added missing block parser registrations

W3D/Rendering:
- Enhanced shadow system (ShadowType, W3DShadowManager)
- Fixed texture manager mipmap/HSV algorithms
- Improved default draw callback parity

Other fixes:
- RAMFile class complete implementation
- StreamingArchiveFile File trait implementation
- Money class xfer() fix
- Weapon defaults, radius damage cone, shared reload time

+++ GeneralsRust/Code/GameEngine/Common/src/common/system/xfer.rs
@@ -14,7 +14,7 @@ use super::snapshot::Snapshot;
 use std::io::{self, Read, Write};
 
 /// Type alias for XferVersion - matches C++ line 29
-pub type XferVersion = u8;
+pub type XferVersion = u32;
 
 /// Xfer mode enumeration - matches C++ Xfer.h lines 33-42
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
